requirements.txt
python-dotenv
streamlit
langchain-groq
crewai